requirements.txt
python-telegram-bot==21.5
fastapi
uvicorn
supabase
python-dotenv
